Move Entity-based components to its own package
