+++ src/com/reco1l/osu/multiplayer/LobbyRoomList.kt
@@ -10,7 +10,7 @@ import com.reco1l.ibancho.data.TeamMode.HEAD_TO_HEAD
 import com.reco1l.ibancho.data.TeamMode.TEAM_VS_TEAM
 import com.reco1l.ibancho.data.WinCondition.*
 import com.reco1l.osu.mainThread
-import com.reco1l.osu.ui.ScrollableList
+import com.reco1l.osu.ui.entity.ScrollableList
 import com.reco1l.toolkt.kotlin.async
 import org.anddev.andengine.entity.sprite.Sprite
 import org.anddev.andengine.entity.text.Text
