+++ src/com/reco1l/osu/ui/entity/StatisticSelector.kt
@@ -1,4 +1,4 @@
-package com.reco1l.osu.ui
+package com.reco1l.osu.ui.entity
 
 import android.opengl.GLES20
 import com.reco1l.ibancho.data.WinCondition.ACCURACY
